requirements.txt
absl-py==2.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Deprecated==1.2.14
fonttools==4.54.1
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.0
namex==0.0.8
numpy==2.1.2
opencv-python==4.10.0.84
optree==0.13.0
packaging==24.1
pillow==11.0.0
pycparser==2.22
pygit2==1.16.0
PyGithub==2.4.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
